project/requirements.txt
langchain==0.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.2
streamlit